Stopped Dropbox synchronisation for the moment and cleaned Dropbox service
