+++ DropboxSync.BLL/IServices/IDropboxService.cs
@@ -8,5 +8,6 @@ namespace DropboxSync.BLL.IServices
 {
     public interface IDropboxService
     {
+        string GeneratedAuthenticationURL();
     }
 }
